chore: reconstruct rag using tabs
